Keep the CI test matrix at one hundred entries
The previous commit added tests/mock_vws/test_model_target_cross_cutting.py
as a new CI matrix entry, but each ci-tests job copies
ci_secrets/vuforia_secrets_${JOB_INDEX}.env from the encrypted secrets
archive, which holds exactly one hundred files. The extra entry pushed
the last job's index to 100, which has no secrets file, so the docs/
job failed before running any tests.

Fold the cross-cutting Model Target tests into
test_model_target_web_api.py, which already has a matrix entry, so the
job count and every job's secrets file stay unchanged.
